FASTAPI/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pydantic
aiofiles
passlib
python-jose
bcrypt